FAudio_internal_simd.c
